Rename the normalize function to avoid name clash
The utils/normalize function used originally by the eb_fret project was
overridden by the new normalize function introduced in MATLAB 2018.
Rename the old function to retain old behavior.

This is found a year ago, but fixed in the imscroll side and never
commited here. The symptom will be an error in the hstep_nw() when using
lsqnonlin fitting. The error was locked down at eb_hmm line 351 by
recognizing the input parameter phi contains NaN.